[#234]fix: return viz를 flowcontrol viz에서 분리
