+++ app/visualize/analysis/stmt/models/flow_control_obj.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field
-from typing import Any
 
 from app.visualize.analysis.stmt.models.stmt_type import StmtType
 
@@ -22,13 +21,4 @@ class PassStmtObj:
 class ContinueStmtObj:
     id: int
     flow_control_type: StmtType = field(default_factory=lambda: StmtType.CONTINUE, init=False)
-    type: StmtType = field(default_factory=lambda: StmtType.FLOW_CONTROL, init=False)
-
-
-@dataclass(frozen=True)
-class ReturnStmtObj:
-    id: int
-    value: Any
-    expr: tuple[str, ...]
-    flow_control_type: StmtType = field(default_factory=lambda: StmtType.RETURN, init=False)
     type: StmtType = field(default_factory=lambda: StmtType.FLOW_CONTROL, init=False)
