+++ app/visualize/generator/converter/flow_control_converter.py
@@ -2,29 +2,15 @@ from app.visualize.analysis.stmt.models.flow_control_obj import (
     BreakStmtObj,
     PassStmtObj,
     ContinueStmtObj,
-    ReturnStmtObj,
 )
-from app.visualize.analysis.stmt.models.stmt_type import StmtType
-from app.visualize.generator.models.flow_control_viz import FlowControlViz, ReturnFlowControlViz
+from app.visualize.generator.models.flow_control_viz import FlowControlViz
 from app.visualize.generator.visualization_manager import VisualizationManager
 
 
 class FlowControlConverter:
 
     @staticmethod
-    def convert(node: PassStmtObj | BreakStmtObj | ContinueStmtObj | ReturnStmtObj, viz_manager: VisualizationManager):
-
-        if node.flow_control_type is StmtType.RETURN:
-            return tuple(
-                ReturnFlowControlViz(
-                    id=node.id,
-                    depth=viz_manager.get_depth(),
-                    returnExpr=expressions,
-                    code=viz_manager.get_code_by_idx(node.id),
-                )
-                for expressions in node.expr
-            )
-
+    def convert(node: PassStmtObj | BreakStmtObj | ContinueStmtObj, viz_manager: VisualizationManager):
         return (
             FlowControlViz(
                 id=node.id,
